backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
opencv-python-headless
supabase
python-dotenv
google-genai

